requirements.txt
google-api-python-client==2.156.0
youtube-transcript-api==0.6.2
google-genai==1.22.0
aiosmtplib==4.0.1
tenacity==9.1.2
aiofiles==24.1.0
markdown==3.7
